🚚 api: script: move repository method and script to course src
