+++ api/src/certification/course/scripts/generate-certification-attestations-by-session-ids.js
@@ -7,25 +7,23 @@ import bluebird from 'bluebird';
 import lodash from 'lodash';
 
 const { isEmpty, compact } = lodash;
-import { logger } from '../../lib/infrastructure/logger.js';
-import * as certificateRepository from '../../lib/infrastructure/repositories/certificate-repository.js';
-import * as certificationCourseRepository from '../../lib/infrastructure/repositories/certification-course-repository.js';
-import * as certificationAttestationPdf from '../../lib/infrastructure/utils/pdf/certification-attestation-pdf.js';
-import { NotFoundError } from '../../lib/domain/errors.js';
-import { learningContentCache as cache } from '../../lib/infrastructure/caches/learning-content-cache.js';
-import { disconnect } from '../../db/knex-database-connection.js';
+import { logger } from '../../../shared/infrastructure/utils/logger.js';
+import * as certificateRepository from '../infrastructure/repositories/certificate-repository.js';
+import * as certificationCourseRepository from '../../../../lib/infrastructure/repositories/certification-course-repository.js';
+import * as certificationAttestationPdf from '../infrastructure/utils/pdf/certification-attestation-pdf.js';
+import { NotFoundError } from '../../../shared/domain/errors.js';
+import { learningContentCache as cache } from '../../../../lib/infrastructure/caches/learning-content-cache.js';
+import { disconnect } from '../../../../db/knex-database-connection.js';
 import * as url from 'url';
 import i18n from 'i18n';
+import { options } from '../../../../lib/infrastructure/plugins/i18n.js';
 import path from 'path';
 const __dirname = url.fileURLToPath(new URL('.', import.meta.url));
 
-const directory = path.resolve(path.join(__dirname, '../../translations'));
+const directory = path.resolve(path.join(__dirname, '../../../../translations'));
 i18n.configure({
-  locales: ['fr', 'en'],
-  defaultLocale: 'fr',
+  ...options,
   directory,
-  objectNotation: true,
-  updateFiles: false,
 });
 
 /**
